logging: Reduce verbosity of some vdsmInterface messages
This patch makes some changes to the logging in vdsmInterface.  First, lower the
verbosity of the VM List message to debug since it is cluttering up the log.
It's only really needed if you are debugging an issue around getting the list of
VMs.

Second, convert the getVmMemoryStats errors so they just return error info in
the HypervisorInterfaceError.  This info will be reported as seen fit by the
caller and there is no need to print it at WARNING verbosity here.  Change the
verbosity of stats_error() to WARNING so that when stats collection errors
happen, the user will see the message the first time the error occurs.

Change-Id: I80701029f45f3e1d791846df302eae13ee3354e7
Signed-off-by: Adam Litke <agl@us.ibm.com>