doc: Update sample config to include GuestBalloon
One of the sample configs has the GuestBalloon collector enabled but the other
does not.  This leads to confusion when users try to load the sample balloon
policy and the necessary collectors are not loaded.  Update the sample config to
help avoid this confusion.

(Note: The long-term solution is to add 'Require' support to the policy so it
can load its own collectors and controllers)

Change-Id: I33012732d2b16d52c631ab23bdebb725f9d00d0f
Signed-off-by: Adam Litke <agl@us.ibm.com>