multi-policy: Set multiple policies at startup
Now that MOM supports multiple policies there should be a way to configure them
to run automatically at startup.  To do this we will make use of a common
concept in other programs, a policy.d directory.  A new configuration option
is added to the config file that allows the user to specify the location of a
policy.d directory from which a set of policies is to be loaded.  Only files
ending in .policy will be considered.  The basename of the file will become the
policy name.

Change-Id: I9230bf6c93cde579ff09739acc7d4c66a6741046
Signed-off-by: Adam Litke <agl@us.ibm.com>