mom: communicate with qemu-guest-agent using hypervisor_iface via libvirt
When the unix socket is defined for the qemu-guest-agent, libvirt connects
to the socket for its internal usage, and so the other client cannot use it.
Instead, libvirt_qemu provides an interface to communicate with the guest
agent. With this patch, GuestQemuAgent will firstly try to communicate with
the guest agent using libvirt via hypervisor_iface. When hypervisor_iface
cannot provide access to the agent, it falls back to the socket interface.

Change-Id: I8ba8d3cda466ca5212f4365ff3ee9ff2864524d6
Signed-off-by: Tomoki Sekiyama <tomoki.sekiyama@hds.com>
Reviewed-on: http://gerrit.ovirt.org/18888
Reviewed-by: Adam Litke <agl@us.ibm.com>
Tested-by: Adam Litke <agl@us.ibm.com>