guest monitor: split monitor from thread
We have the GuestMonitor class which inherits from threading.Thread,
and takes care of the per-VM monitor book-keeping and of the monitoring
loop.

This patch splits the class in two, with an explicit monitoring thread,
decoupling the two responsabilities.

This is needed for better scalability, because this way we can now
optionally demand the actual invocation of the monitoring code to
the hypervisor, which can integrate it with its monitoring loop(s).

Moreover, this patch also adds explicit registration and deregistration
methods for Guest Monitor objects inside the Guest Manager.

Besides a little cleanup, this patch makes room for
a future patch want to optionally demand the collection of
the MOM stats to hypervisor.

The old behaviour is preserved through fallback code.

Change-Id: I224104f8b353d92a420723d3415df94aaf68da0c
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/31591
Continuous-Integration: Jenkins CI
Reviewed-by: Martin Sivák <msivak@redhat.com>