Monitor should not crash when unexpected exception happens
The monitor thread should report the crash to log, behave as
if nothing was collected and try again in the next cycle.

The unexpected exception might be triggered for example by
a freshly finished migration. The guest monitor reads the /proc
files that belong to the migrated process and they disappear
as a result of qemu performing the post-migration cleanup.

Change-Id: I318c55a3215df8ccaf4b0e2d16d1d9a3b06f9d1c
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1124595
Signed-off-by: Martin Sivak <msivak@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/32030
Reviewed-by: Adam Litke <alitke@redhat.com>