guest manager: single-thread monitoring
So far MOM used one thread per Guest Monitor.
The two main reasons were
- to minimize interference between VMs
- the lack of a suitable bulk API (especially from libvirt).

However, when running inside VDSM, MOM can leverage
the safety which VDSM provides, and its bulk APIs.
Moreover, libvirt recently gained bulk APIs as well.

This patch changes the internals of Guest Manager to
fully decouple Guest Monitor and threads.
It is still possible to run one thread per Guest Monitor
as before, but it is now possible to mix and match
threads and monitors with complete freedom.

This change paves the road for better and more efficient
integration with VDSM. In that case, we want to be able
to run all the MOM monitoring with just one thread.

Change-Id: I96b96fef0c0c42fd45b76754da9c629865ab1954
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/37826
Continuous-Integration: Jenkins CI
Reviewed-by: Martin Sivák <msivak@redhat.com>