New hypervisor interface for remote VDSM over XML-RPC with stats caching
This patch adds a new vdsmxmlrpcInterface that uses getAllVmStats
for retrieving all information about all VMs at the same time.

The stats structure is then cached for 5 seconds to make sure we
are not overloading VDSM with constant requests.

XML-RPC client is stateless so it does not matter if VDSM dies or
restarts between requests, MOM will treat the situation as if the
host was empty and will keep running.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1227714
Change-Id: I314f39f9020ee257827145a57f585e7921a913e3
Signed-off-by: Martin Sivak <msivak@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/41570
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>