Allow using unix socket for the RPC port
This adds a code that is able to start an XML-RPC server listening
on unix socket. MOM first tries to see if the port can be parsed as
a number and uses it as a path when it is not.

port -1 - no RPC
port <number> - TCP based XML-RPC
port <anything else> - unixsock based XML-RPC

Change-Id: I6617d92c94259dff604acaa2b6481ef9595522a4
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1227714
Signed-off-by: Martin Sivak <msivak@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/42228
Reviewed-by: Adam Litke <alitke@redhat.com>