vdsm xml rpc: fix locking
Multiple threads will call concurrently the hypervisor interface.
Without proper locking, we will have races. The most noticeable
effect is an unbounded amount of calls to VDSM, due to the raceful
age check. This completely defeats the purpose of caching.

This patch adds a lock to make sure the check is safe, and that
the caching is effective.

Change-Id: I8a6d058ee2631bd100153c2091400a2a35b4fb1d
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/42574
Reviewed-by: Martin Sivák <msivak@redhat.com>
Continuous-Integration: Jenkins CI