Improve the message when getVmMemoryStats() returns 0
Mom is currently expecting a non-zero return for memory usage to assume
ovirt-guest-agent is active, and throws an warning log along with an 'error'
message if the first check has a 0ed return.

This patch removes the 'error' string in the warning message and also
includes the VM uuid in the message to improve the message relevance.

Change-Id: I4afba631f190440810b3f022987ff152773ac6e6
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1234953
Signed-off-by: Amador Pahim <apahim@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/45596
Reviewed-by: Adam Litke <alitke@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Martin Sivák <msivak@redhat.com>