Fix a possible deadlock in python 3
The deadlock can happen when a VM is removed.
The _check_guest_monitors method in GuestManager removes
the monitor from a dictionary while iterating over it, which
raises RuntimeException and a lock is never be released.

Later when VDSM asks MOM for KSM statistics, the RPC handler
tries to acquire the lock but never succeeds. This causes
deadlock, and the RPC server cannot be shutdown as it is waiting
for the lock.

Change-Id: Ia200960d9cb91e688d10f773ee9b027af5cb214c
Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/102714
Continuous-Integration: Jenkins CI <jenkins@ovirt.org>
Tested-by: Michal Skrivanek <michal.skrivanek@redhat.com>
Reviewed-by: Milan Zamazal <mzamazal@redhat.com>